backend/pyproject.toml
[project]
name = "document-master-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "openvino>=2024.6.0",
    "openvino-dev[onnx,pytorch,tools]>=2024.6.0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "ruff>=0.14.5",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]
